refactor(feature/donkey/#13): style 파일 이름 camel case로 변경
