+++ instagram/src/components/UserPage/UserHeader.jsx
@@ -1,6 +1,6 @@
 import ProfilePic from "./ProfilePic";
 import HeaderSection from "./HeaderSection";
-import { HeaderContainer } from "../../styles/UserPage/UserHeader_style";
+import { HeaderContainer } from "../../styles/UserPage/UserHeaderStyle";
 
 function UserHeader() {
   return (
