+++ instagram/src/styles/ProfileEditPage/DetailsFormStyle.jsx
