+++ instagram/src/styles/UserPage/HeaderSectionStyle.jsx
