+++ instagram/src/styles/UserPage/UserPageStyle.jsx
